Erste Fixes für CEGUI 0.3   more to come soon (ne Menge :))
git-svn-id: svn://svn.berlios.de/dsa-hl@826 4c79e8ff-cfd4-0310-af45-a38c79f83013

+++ rllooks/trunk/RastullahLook/src/RLComboEditbox.cpp
@@ -85,7 +85,7 @@ RLComboEditbox::~RLComboEditbox(void)
     Return the text code point index that is rendered closest to screen
     position 'pt'.  
 *************************************************************************/
-ulong RLComboEditbox::getTextIndexFromPosition(const Point& pt) const
+size_t RLComboEditbox::getTextIndexFromPosition(const Point& pt) const
 {
     //
     // calculate final window position to be checked
